@@ -1,7 +1,8 @@
 {
   "name": "@kevinastone/eslint-config",
-  "version": "0.3.0",
+  "version": "0.4.0",
   "description": "",
+  "type": "module",
   "main": "index.js",
   "bin": {
     "pnp-eslint": "./bin/pnp-eslint.js"
@@ -24,25 +25,27 @@
   },
   "homepage": "https://github.com/kevinastone/eslint-config#readme",
   "dependencies": {
-    "@typescript-eslint/eslint-plugin": "^2.26.0",
-    "@typescript-eslint/parser": "^2.26.0",
-    "eslint": "^6.8.0",
-    "eslint-config-airbnb-typescript": "^7.2.0",
-    "eslint-config-prettier": "^6.0.0",
-    "eslint-plugin-eslint-comments": "^3.1.2",
-    "eslint-plugin-import": "^2.20.2",
-    "eslint-plugin-jest": "^23.1.1",
-    "eslint-plugin-prettier": "^3.1.0",
-    "eslint-plugin-react": "^7.19.0",
-    "eslint-plugin-react-hooks": "^3.0.0",
-    "eslint-plugin-unicorn": "^18.0.1",
-    "typescript": ">=4.0"
+    "@eslint/js": "^9.9.0",
+    "@typescript-eslint/eslint-plugin": "^8.1.0",
+    "@typescript-eslint/parser": "^8.1.0",
+    "eslint": "^9.9.0",
+    "eslint-config-airbnb-typescript": "^18.0.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jest": "^28.8.0",
+    "eslint-plugin-prettier": "^5.2.1",
+    "eslint-plugin-react": "^7.35.0",
+    "eslint-plugin-react-hooks": "^4.6.2",
+    "eslint-plugin-unicorn": "^55.0.0",
+    "typescript": "^5.5.4",
+    "typescript-eslint": "^8.1.0"
   },
   "peerDependencies": {
-    "eslint": "^6.8.0",
+    "eslint": "^9.9.0",
     "typescript": "*"
   },
   "devDependencies": {
-    "prettier": "^2.0.2"
+    "prettier": "^3.3.3",
+    "@types/eslint__js": "^8.42.3"
   }
 }
