@@ -1,3 +1,9 @@
-module.exports = {
-  extends: ['./shared', './react'],
-};
+import react from './react.js';
+import shared from './shared.js';
+// import jest from './jest.js';
+
+export default [
+  ...shared,
+  // ...jest,
+  ...react,
+];
