@@ -1,12 +1,21 @@
-module.exports = {
-  plugins: ['react', 'react-hooks'],
-  extends: ['plugin:react/recommended', 'plugin:react-hooks/recommended'],
-  settings: {
-    react: {
-      version: 'detect',
+import react from 'eslint-plugin-react';
+// import reactHooks from 'eslint-plugin-react-hooks';
+import tseslint from 'typescript-eslint';
+
+export default tseslint.config(
+  react.configs.flat.recommended,
+  // https://github.com/facebook/react/issues/28313
+  // reactHooks.configs.flat.recommended,
+  {
+    rules: {
+      'react/display-name': 'off',
     },
   },
-  rules: {
-    'react/display-name': 'off',
-  },
-};
+  {
+    settings: {
+      react: {
+        version: "detect",
+      }
+    }
+  }
+);
